Update layout & code
Margins fixed.
Fixed several logical bugs in code.
Update the preview page and also merged it with the page no button selected, that big title is removed.
The preview page now looked like the server browser and supports sorting by label, x,y,w,h.
The total width of the editor is increased.
A new tab Preview is created.
The preview all buttons function is moved to the preview tab.